requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.1.8

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Document Processing
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4

# Embeddings
sentence-transformers==2.3.1
torch==2.1.2
numpy==1.26.3

# LLM - Google Gemini
google-generativeai==0.4.1

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# UUID
uuid==1.30
